-ViewModel de empleos. -Botón de Ver Más muestra la información del empleo en un form de readonly. -Funcionalidad de empleo inteligente lista. -Correcciones en el insertar empleo para tomar el correo y la compañía de la base de datos.

+++ BackEnd/DAL/IDALGenerico.cs
@@ -7,7 +7,7 @@ namespace BackEnd.DAL
 {
     public interface IDALGenerico<TEntity> where TEntity : class
     {
-        TEntity Get(int id);
+        TEntity Get(long id);
         IEnumerable<TEntity> GetAll();
         IEnumerable<TEntity> Find(Expression<Func<TEntity, bool>> predicate);
